@@ -3,17 +3,19 @@
   "version": "2.3.0",
   "private": false,
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@material-ui/icons": "^4.9.1",
-    "apexcharts": "^3.21.0",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/material": "^5.10.15",
+    "@mui/styles": "^5.10.15",
+    "apexcharts": "^3.36.3",
     "moment": "^2.29.1",
-    "react": "^16.13.1",
+    "react": "^18.2.0",
     "react-apexcharts": "^1.3.7",
-    "react-dom": "^16.13.1",
+    "react-dom": "^18.2.0",
     "wn8-color": "^1.0.0"
   },
   "devDependencies": {
-    "react-scripts": "^5.0.0"
+    "react-scripts": "^5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
